requirements.txt
Django==2.2.23
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
entrypoints==0.3
packaging==20.9
pandas==1.2.4
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1

